Move module version checker to serialization.py
