+++ torch/_utils_internal.py
@@ -2,7 +2,6 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import os
 import inspect
-import warnings
 
 # this arbitrary-looking assortment of functionality is provided here
 # to have a central place for overrideable behavior. The motivating
@@ -51,43 +50,6 @@ def get_source_lines_and_file(obj):
             "Make sure original .py files are available. Original error: {}").format(filename, e))
 
     return sourcelines, file_lineno, filename
-
-
-def check_module_version_greater_or_equal(module, req_version_tuple, error_if_malformed=True):
-    '''
-    Check if a module's version satisfies requirements
-
-    Usually, a module's version string will be like 'x.y.z', which would be represented
-    as a tuple (x, y, z), but sometimes it could be an unexpected format. If the version
-    string does not match the given tuple's format up to the length of the tuple, then
-    error and exit or emit a warning.
-
-    Args:
-        module: the module to check the version of
-        req_version_tuple: tuple (usually of ints) representing the required version
-        error_if_malformed: whether we should exit if module version string is malformed
-
-    Returns:
-        requirement_is_met: bool
-    '''
-    try:
-        version_strs = module.__version__.split('.')
-        # Cast module version fields to match the types of the required version
-        module_version = tuple(
-            type(req_field)(version_strs[idx]) for idx, req_field in enumerate(req_version_tuple)
-        )
-        requirement_is_met = module_version >= req_version_tuple
-
-    except:
-        message = ("'%s' module version string is malformed '%s' and cannot be compared"
-            " with tuple %s" % (module.__name__, module.__version__, str(req_version_tuple)))
-        if error_if_malformed:
-            raise Exception(message)
-        else:
-            warnings.warn(message + ', but continuing assuming that requirement is met')
-            requirement_is_met = True
-
-    return requirement_is_met
 
 
 TEST_MASTER_ADDR = '127.0.0.1'
